Move the hierarchy adjustment functions into beats helpers
