+++ v2/helpers/beats.js
@@ -117,3 +117,56 @@ export function maxDepth(beats) {
     Number.NEGATIVE_INFINITY
   )
 }
+
+export const addLevelToHierarchy = (currentTree, newParentId, bookId) => {
+  const node = {
+    autoOutlineSort: true,
+    bookId: bookId,
+    fromTemplateId: null,
+    id: newParentId,
+    position: 0,
+    time: 0,
+    title: 'auto',
+    expanded: true,
+  }
+  const withNewParent = tree.addNode('id')(currentTree, null, node)
+  const newTopLevelNodes = tree.children(withNewParent, null)
+  const newParent = newTopLevelNodes.reduce((highestIdTopLevelNode, nextTopLevelNode) => {
+    if (highestIdTopLevelNode.id > nextTopLevelNode.id) return highestIdTopLevelNode
+    else return nextTopLevelNode
+  }, newTopLevelNodes[0])
+  const oldTopLevelNodes = tree.children(currentTree, null)
+  const newTree = oldTopLevelNodes.reduce((accTree, { id }) => {
+    return tree.moveNode(accTree, id, newParent.id)
+  }, withNewParent)
+  return newTree
+}
+
+export const removeLevelFromHierarchy = (currentTree) => {
+  const topLevelNodesToRemove = tree.children(currentTree, null)
+  const nodesToMoveToTopLevel = topLevelNodesToRemove.flatMap(({ id }) => {
+    return tree.children(currentTree, id)
+  })
+  const withNodesMoved = nodesToMoveToTopLevel.reduce((newTree, { id }) => {
+    return tree.moveNode(newTree, id, null)
+  }, currentTree)
+  const newTree = topLevelNodesToRemove.reduce((newTree, { id }) => {
+    return tree.deleteNode(newTree, id)
+  }, withNodesMoved)
+  return newTree
+}
+
+export const adjustHierarchyLevels = (targetHierarchyDepth) => (
+  currentTree,
+  nextParentId,
+  bookId
+) => {
+  const maximumDepth = maxDepth(currentTree)
+  if (targetHierarchyDepth === maximumDepth) {
+    return currentTree
+  } else if (targetHierarchyDepth > maximumDepth) {
+    return addLevelToHierarchy(currentTree, nextParentId, bookId)
+  } else {
+    return removeLevelFromHierarchy(currentTree)
+  }
+}
