+++ v2/reducers/beats.js
@@ -19,20 +19,10 @@ import {
 } from '../constants/ActionTypes'
 import { beat as defaultBeat } from '../store/initialState'
 import { newFileBeats } from '../store/newFileState'
-import { positionReset, nextPositionInBook, moveNextToSibling, maxDepth } from '../helpers/beats'
+import { positionReset, nextPositionInBook, moveNextToSibling } from '../helpers/beats'
 import { associateWithBroadestScope } from '../helpers/lines'
-import {
-  addNode,
-  deleteNode,
-  editNode,
-  nodeParent,
-  filter,
-  newTree,
-  findNode,
-  children,
-  moveNode,
-} from './tree'
-import { nextId } from '../helpers/beats'
+import { addNode, deleteNode, editNode, nodeParent, filter, newTree, findNode } from './tree'
+import { nextId, adjustHierarchyLevels } from '../helpers/beats'
 
 // bookId is:
 // Union of:
@@ -57,55 +47,6 @@ const addNodeToState = (state, bookId, position, title, parentId) => {
   return {
     ...state,
     [bookId]: add(state[bookId], parentId, node),
-  }
-}
-
-const addLevelToHierarchy = (currentTree, newParentId, bookId) => {
-  const node = {
-    autoOutlineSort: true,
-    bookId: bookId,
-    fromTemplateId: null,
-    id: newParentId,
-    position: 0,
-    time: 0,
-    title: 'auto',
-    expanded: true,
-  }
-  const withNewParent = add(currentTree, null, node)
-  const newTopLevelNodes = children(withNewParent, null)
-  const newParent = newTopLevelNodes.reduce((highestIdTopLevelNode, nextTopLevelNode) => {
-    if (highestIdTopLevelNode.id > nextTopLevelNode.id) return highestIdTopLevelNode
-    else return nextTopLevelNode
-  }, newTopLevelNodes[0])
-  const oldTopLevelNodes = children(currentTree, null)
-  const newTree = oldTopLevelNodes.reduce((accTree, { id }) => {
-    return moveNode(accTree, id, newParent.id)
-  }, withNewParent)
-  return newTree
-}
-
-const removeLevelFromHierarchy = (currentTree) => {
-  const topLevelNodesToRemove = children(currentTree, null)
-  const nodesToMoveToTopLevel = topLevelNodesToRemove.flatMap(({ id }) => {
-    return children(currentTree, id)
-  })
-  const withNodesMoved = nodesToMoveToTopLevel.reduce((newTree, { id }) => {
-    return moveNode(newTree, id, null)
-  }, currentTree)
-  const newTree = topLevelNodesToRemove.reduce((newTree, { id }) => {
-    return deleteNode(newTree, id)
-  }, withNodesMoved)
-  return newTree
-}
-
-const adjustHierarchyLevels = (targetHierarchyDepth) => (currentTree, nextParentId, bookId) => {
-  const maximumDepth = maxDepth(currentTree)
-  if (targetHierarchyDepth === maximumDepth) {
-    return currentTree
-  } else if (targetHierarchyDepth > maximumDepth) {
-    return addLevelToHierarchy(currentTree, nextParentId, bookId)
-  } else {
-    return removeLevelFromHierarchy(currentTree)
   }
 }
